🔥 The exercise 07
This one was harder than I expected. Actually I am not sure If this structure is usable for 2 or 3 removals. I think you should change the distance you can see problems in order to increase the removal count. But I was amazed that GPT4 made a better solution than most of the users